+++ docs/src/components/CourseBanner/index.tsx
@@ -1,0 +1,2 @@
+export {default} from './CourseBanner';
+export {default as CourseVideoBanner} from './CourseVideoBanner';