+++ examples/example-app-router-mixed-routing/src/app/(public)/[locale]/about/page.tsx
@@ -3,15 +3,11 @@ import {setRequestLocale} from 'next-intl/server';
 import {use} from 'react';
 import PageTitle from '@/components/PageTitle';
 
-type Props = {
-  params: Promise<{locale: Locale}>;
-};
-
-export default function About({params}: Props) {
+export default function About({params}: PageProps<'/[locale]/about'>) {
   const {locale} = use(params);
 
   // Enable static rendering
-  setRequestLocale(locale);
+  setRequestLocale(locale as Locale);
 
   const t = useTranslations('About');
   return <PageTitle>{t('title')}</PageTitle>;