+++ examples/example-app-router-playground/src/app/[locale]/news/[articleId]/page.tsx
@@ -3,14 +3,9 @@ import {Locale, useTranslations} from 'next-intl';
 import {use} from 'react';
 import {getPathname} from '@/i18n/navigation';
 
-type Props = {
-  params: Promise<{
-    locale: Locale;
-    articleId: string;
-  }>;
-};
-
-export async function generateMetadata({params}: Props): Promise<Metadata> {
+export async function generateMetadata({
+  params
+}: PageProps<'/[locale]/news/[articleId]'>): Promise<Metadata> {
   const {locale, articleId} = await params;
 
   return {
@@ -20,13 +15,15 @@ export async function generateMetadata({params}: Props): Promise<Metadata> {
           pathname: '/news/[articleId]',
           params: {articleId}
         },
-        locale
+        locale: locale as Locale
       })
     }
   };
 }
 
-export default function NewsArticle(props: Props) {
+export default function NewsArticle(
+  props: PageProps<'/[locale]/news/[articleId]'>
+) {
   const {articleId} = use(props.params);
   const t = useTranslations('NewsArticle');
   return <h1>{t('title', {articleId})}</h1>;