+++ examples/example-app-router/src/app/[locale]/pathnames/page.tsx
@@ -3,15 +3,13 @@ import {setRequestLocale} from 'next-intl/server';
 import {use} from 'react';
 import PageLayout from '@/components/PageLayout';
 
-type Props = {
-  params: Promise<{locale: Locale}>;
-};
-
-export default function PathnamesPage({params}: Props) {
+export default function PathnamesPage({
+  params
+}: PageProps<'/[locale]/pathnames'>) {
   const {locale} = use(params);
 
   // Enable static rendering
-  setRequestLocale(locale);
+  setRequestLocale(locale as Locale);
 
   const t = useTranslations('PathnamesPage');